# Experimentations to make sortperm faster