It has been noted at https://github.com/JuliaLang/julia/issues/939 that `sortperm` is much slower than `sort` and than some implementations in other languages. In particular, it is much slower than `forderv` from the data.table R package, even when using Radix sort as data.table does.

In this PR I've done a few experiments to improve this. When vectors get very large (> 1M entries),  the performance gap between `sort` (which is roughly as fast as data.table with radix sort) and `sortperm` increases dramatically, probably due to cache locality issues: since the original data isn't sorted as the same time as indices, memory accesses are irregular. With some changes so that both the permutation index and the data are sorted at the same time,  the ratio between `sort` and `sortperm` timings can be reduced to about 2 (instead of exploding as currently). Unfortunately, this PR also slows down `sortperm` for smaller vectors (below 100k entries):
```julia
julia> using SortingAlgorithms, BenchmarkTools


## 100k entries
julia> x = rand(Int, 100_000);

julia> @btime sort(x, alg=RadixSort);
  5.135 ms (18 allocations: 1.72 MiB)

# Before
julia> @btime sortperm(x, alg=RadixSort);
  6.531 ms (19 allocations: 1.72 MiB)

# After commit 1
julia> @btime sortperm(x, alg=RadixSort);
  9.459 ms (23 allocations: 2.53 MiB)

# After commit 2
julia> @btime sortperm(x, alg=RadixSort);
  8.082 ms (23 allocations: 3.24 MiB)


## 1M entries
julia> x = rand(Int, 1_000_000);

julia> @btime sort(x, alg=RadixSort);
  56.841 ms (18 allocations: 15.45 MiB)

# Before
julia> @btime sortperm(x, alg=RadixSort);
  437.113 ms (19 allocations: 15.45 MiB)

# After commit 1
julia> @btime sortperm(x, alg=RadixSort);
  113.574 ms (23 allocations: 23.12 MiB)

# After commit 2
julia> @btime sortperm(x, alg=RadixSort);
  98.436 ms (23 allocations: 30.71 MiB)


## 10M entries
julia> x = rand(Int, 10_000_000);

julia> @btime sort(x, alg=RadixSort);
  723.494 ms (18 allocations: 152.78 MiB)

# Before
julia> @btime sortperm(x, alg=RadixSort);
  6.239 s (19 allocations: 152.78 MiB)

# After commit 1
julia> @btime sortperm(x, alg=RadixSort);
  1.335 s (23 allocations: 229.12 MiB)

# After commit 2
julia> @btime sortperm(x, alg=RadixSort);
  1.306 s (23 allocations: 305.37 MiB)
```

Note the two commits: the first one implements a cleaner and more generic approach based on a `PermVector` wrapper, which could easily be adapted to other algorithms. But is significantly slower than the second one (suggestions welcome).