@JeffreySarnoff 2019-11-17
> Unfortunately, this PR also slows down sortperm for smaller vectors (below 100k entries)
```
sortperm(x, _etc_) = length(x) < 100_000 ? sortperm_smaller(x, _etc_) : sortperm_larger(x, _etc_)
```
trait-based off e.g. `abstract type ShortVectorToSort{T} <: Vector{T} end` if that helps