@nalimilan 2019-11-17
> Do we have test cases for non-random inputs? It perhaps begs the question what other cases should one test on.

Yes, clearly we need more benchmarks, especially if we have to identify a size threshold.

> > Unfortunately, this PR also slows down sortperm for smaller vectors (below 100k entries)
> 
> ```
> sortperm(x, _etc_) = length(x) < 100_000 ? sortperm_smaller(x, _etc_) : sortperm_larger(x, _etc_)
> ```
> 
> trait-based off e.g. `abstract type ShortVectorToSort{T} <: Vector{T} end` if that helps

Sure, taking two different methods based on a threshold is always possible. But it's not great in terms of code simplicity.

I should have noted that somehow data.table's `forderv` is still faster than `sortperm` even with this PR (by about 15% for the 1M case, and 30% for the 10M case).