@JeffreySarnoff 2019-11-17
_it's not great in terms of code simplicity (agreed)_

My note presupposed that this is a widely used and for certain classes of application, unusually performance-sensitive function.  The estimated algorithmic split length appears sufficiently high that a sizeable count of uses of the function would incur that introduced slowness.  Simultaneously, the import of the introduced fastness for very large vectors is similarly significant.  I am not aware of an internally adaptive approach that is relatively lightweight.  There certainly may be ..