@xiaodaigh 2019-12-01
For reference, I have implemented faster `sortperm` in the form of `fsortperm` in [SortingLab.jl](https://github.com/xiaodaigh/SortingLab.jl).

The performance difference is stark for `Vector{Union{Missing, T}}`, see https://discourse.julialang.org/t/ann-sortinglab-jl-0-2-2-5x-faster-sort-and-sort-perm-for-vector-union-missing-t/31704