@nalimilan 2019-12-01
Thanks. I've just tried that, and for `Vector{Int}` it seems to be about as fast as this PR for very large inputs, but much slower than this PR (and therefore than the current state) for smaller inputs. Though they have the same performance if the number of unique values is small (100). Actually, even the current state is as fast as this PR and as `fsortperm` in that case, disregarding the vector length.
```julia
julia> x = rand(-1000_000:1000_000, 10_000);

# Before
julia> @btime sortperm(x, alg=RadixSort);
  395.240 μs (19 allocations: 350.00 KiB)

# Second commit in this PR
julia> @btime sortperm(x, alg=RadixSort);
  403.349 μs (23 allocations: 506.41 KiB)

julia> @btime fsortperm(x);
  953.816 μs (19 allocations: 2.31 MiB)

julia> x = rand(-1000_000:1000_000, 100_000);

# Before
julia> @btime sortperm(x, alg=RadixSort);
  4.356 ms (19 allocations: 1.72 MiB)

# Second commit in this PR
julia> @btime sortperm(x, alg=RadixSort);
  4.607 ms (23 allocations: 3.24 MiB)

julia> @btime fsortperm(x);
  4.982 ms (19 allocations: 5.05 MiB)

julia> x = rand(-1000_000:1000_000, 1000_000);

# Before
julia> @btime sortperm(x, alg=RadixSort);
  272.243 ms (19 allocations: 15.45 MiB)

# Second commit in this PR
julia> @btime sortperm(x, alg=RadixSort);
  54.644 ms (23 allocations: 30.71 MiB)

julia> @btime fsortperm(x);
  56.045 ms (19 allocations: 32.52 MiB)

julia> x = rand(-10_000_000:10_000_000, 10_000_000);

# Before
julia> @btime sortperm(x, alg=RadixSort);
  7.812 s (19 allocations: 152.78 MiB)

# Second commit in this PR
julia> @btime sortperm(x, alg=RadixSort);
  735.583 ms (23 allocations: 305.37 MiB)

julia> @btime fsortperm(x);
  682.264 ms (19 allocations: 307.18 MiB)

```

(BTW, I had to use `-1000_000:1000_000` for `fsortperm` to work since there's a bug due to overflow when checking the range with `rand(Int, n)`.)

Can you explain what's the approach taken by `fsortperm` in a few words?