@xiaodaigh 2019-12-01
> BTW, I had to use `-1000_000:1000_000` for `fsortperm` to work since there's a bug due to overflow when checking the range with `rand(Int, n)`

Oh, there is a bug? I might look at it closely later. But please submit a bug report if you find more bugs.

> Can you explain what's the approach taken by `fsortperm` in a few words?
For most types, `fsortperm` works by modifying the radix algorithm so that it sorts the original vector but also rearranges an index at the same time. In effect, it's a more efficient version of this

```julia
fsortperm(x) = begin
  # not the actual implementation of `fsortperm`, just an illustration
  x_and_index = collect(zip(x, 1:length(x)))
  sort!(x_and_index, by = x->x[1]) # sort by x, but index gets shuffled around too
  map(x->x[2], x_and_index) # return the index only
end
```

Actually, the original use case, I had was I wanted to sort the column and carry an index as well, so that after sorting it, I have the sorted column and a sortperm index. So I can rearrange DataFrames more efficiently this way.

The real power of `fsortperm` comes from being able to sort `Vector{Union{T, Missing}}` as well