@xiaodaigh 2019-12-01
> > Oh, there is a bug? I might look at it closely later. But please submit a bug report if you find more bugs.
> 
> Yes I think you need to handle overflow (see how `Base.sort` does this).
> 
> > For most types, `fsortperm` works by modifying the radix algorithm so that it sorts the original vector but also rearranges an index at the same time. In effect, it's a more efficient version of this
> 
> OK, so that's essentially what this PR does, right?

Sounds like it